refactor(contact): extract contact form to separate component
