+++ src/components/contact/contact-form/contact-form.component.scss
@@ -1,0 +1,10 @@
+mat-form-field {
+  width: 100%;
+  margin-bottom: 16px;
+}
+
+button[disabled] {
+  background-color: var(--mat-sys-primary);
+  color: var(--mat-sys-on-primary);
+  opacity: 0.6;
+}
