+++ src/components/contact/contact.component.scss
@@ -3,11 +3,6 @@ mat-card {
   box-shadow: none;
 }
 
-mat-form-field {
-  width: 100%;
-  margin-bottom: 16px;
-}
-
 .form-container {
   display: flex;
   flex-direction: column;
@@ -34,10 +29,4 @@ mat-form-field {
   height: 40px;
   width: 40px;
   margin-right: 16px;
-}
-
-button[disabled] {
-  background-color: var(--mat-sys-primary);
-  color:  var(--mat-sys-on-primary);
-  opacity: 0.6;
 }
