+++ src/components/contact/contact.component.ts
@@ -1,48 +1,21 @@
 import { Component } from '@angular/core';
-import {
-  FormGroup,
-  FormControl,
-  ReactiveFormsModule,
-  Validators,
-} from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
 import { MatCardModule } from '@angular/material/card';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatDividerModule } from '@angular/material/divider';
 
+import { ContactFormComponent } from './contact-form/contact-form.component';
+
 @Component({
   selector: 'mdlv-contact',
   imports: [
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatButtonModule,
     MatIconModule,
     MatCardModule,
     MatGridListModule,
     MatDividerModule,
+    ContactFormComponent,
   ],
   templateUrl: './contact.component.html',
   styleUrl: './contact.component.scss',
 })
-export class ContactComponent {
-  //TODO: Add privacy policy checkbox
-  contactForm = new FormGroup({
-    name: new FormControl('', Validators.required),
-    email: new FormControl('', [Validators.required, Validators.email]),
-    phone: new FormControl('', Validators.required),
-    message: new FormControl('', Validators.required),
-  });
-
-  onSubmit() {
-    console.warn(this.contactForm.value);
-    // TODO: Implement solution to send email; prob use EmailJS
-  }
-
-  get requiredErrorMessage() {
-    return 'To pole jest wymagane';
-  }
-}
+export class ContactComponent { }
